+++ src/components/common/Nav/index.tsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import AuthNav from './AuthNav';
+import CategoryNav from './CategoryNav';
+
+const Nav = () => {
+  return (
+    <>
+      <CategoryNav />
+      <AuthNav />
+    </>
+  );
+};
+
+export default Nav;
